projects/italian-logistics-ai/nexus-engine/nexus-backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
requests==2.32.3
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
googlemaps==4.10.0
twilio==9.0.4
python-dotenv==1.0.1
openai==1.63.2
python-multipart==0.0.20
